Add code to work without local PDFs

+++ manual_check.py
@@ -3,15 +3,43 @@ seen_files = set()
 
 import re
 import signal
-import sqlite3
 import subprocess
 import os
 import math
-from PyPDF2 import PdfReader
 
-con = sqlite3.connect("devices.db")
-cur = con.cursor()
+import argparse
 
+parser = argparse.ArgumentParser(
+    prog="Manual PDF Check",
+    description="Semi-automates predicate device data entry.",
+    epilog="Stay cool kiddo",
+)
+
+parser.add_argument(
+    "-l",
+    "--local_pdfs",
+    default=False,
+    action=argparse.BooleanOptionalAction,
+)
+parser.add_argument(
+    "-d",
+    "--directory",
+    default="/home/william/Downloads/fda-pdfs-ocr/scraper-combined",
+    action="store",
+)
+
+parser.add_argument(
+    "-r",
+    "--reversed",
+    default=False,
+    action=argparse.BooleanOptionalAction,
+)
+
+args = parser.parse_args()
+
+DOWNLOAD_PATH = args.directory
+
+# read in the old data
 with open("manually_added_links.csv", "r") as f:
     data = f.readlines()
     for line in data:
@@ -19,50 +47,24 @@ with open("manually_added_links.csv", "r") as f:
         seen_files.add(target)
 
 
-res = cur.execute("SELECT k_number FROM device WHERE statement_or_summary = 'Summary' AND k_number NOT IN (SELECT k_number FROM device JOIN predicate_graph_edge ON device.k_number = predicate_graph_edge.node_to);")
-missing_edges = res.fetchall()
-
-DOWNLOAD_PATH = "/home/wcedmisten/Downloads/fda-pdfs-ocr/scraper-combined"
-
 def good_input(input_str):
     return input_str == "F" or input_str == "S" or re.match("K\d{6}", input_str)
 
-def check_for_predicate_description(filename):
-    if os.path.isfile(filename + ".txt"):
-        with open (filename + ".txt", "r") as f:
-            lines = f.read()
-            if "equivalent" in lines or "predicate" in lines or "equivalence" in lines or "equivalency" in lines:
-                return
-    else:
-        try:
-            with open(filename, "rb") as f:
-                pdf = PdfReader(f)
-
-                for p in range(len(pdf.pages)):
-                    page = pdf.pages[p]
-                    try:
-                        lines= page.extract_text()
-                        if "equivalent" in lines or "predicate" in lines or "equivalence" in lines or "equivalency" in lines:
-                            return
-                    except Exception as e:
-                        print("Could not parse PDF page")
-                        print(e)
-
-        except Exception as e:
-            print("Could not open PDF file")
 
 def process_pdf(k_number):
     filename = f"{DOWNLOAD_PATH}/{k_number}.pdf"
-    if not os.path.isfile(filename):
+    print(filename)
+    if not os.path.isfile(filename) and args.local_pdfs:
         print("Could not find PDF, skipping.")
         data.append(f"{k_number},S\n")
         return
 
-    process = subprocess.Popen(["evince", filename], shell=False)
+    if args.local_pdfs:
+        process = subprocess.Popen(["evince", filename], shell=False)
 
-    check_for_predicate_description(filename)
-
-    print(f"https://www.accessdata.fda.gov/scripts/cdrh/cfdocs/cfpmn/pmn.cfm?ID={k_number}")
+    print(
+        f"https://www.accessdata.fda.gov/scripts/cdrh/cfdocs/cfpmn/pmn.cfm?ID={k_number}"
+    )
     print()
 
     prompt = f"Enter a predicate ID for {k_number}, or press F to finish inputting predicates for this device. Press S if no predicates can be found.\n"
@@ -74,15 +76,17 @@ def process_pdf(k_number):
         response = response.replace("O", "0").replace(" ", "")
         if good_input(response):
             if response == "F":
-                process.send_signal(signal.SIGTERM)
-                process.terminate()
-                process.kill()
+                if args.local_pdfs:
+                    process.send_signal(signal.SIGTERM)
+                    process.terminate()
+                    process.kill()
                 return
             if response == "S":
                 data.append(f"{k_number},S\n")
-                process.send_signal(signal.SIGTERM)
-                process.terminate()
-                process.kill()
+                if args.local_pdfs:
+                    process.send_signal(signal.SIGTERM)
+                    process.terminate()
+                    process.kill()
                 return
             elif re.match("K\d{6}", response):
                 data.append(f"{k_number},{response}\n")
@@ -91,17 +95,28 @@ def process_pdf(k_number):
             print("Unexpected response, try again.")
             response = input(prompt)
 
+
+with open("missing_predicates.txt", "r") as f:
+    missing_edges = f.readlines()
+    if args.reversed:
+        missing_edges.reverse()
+
 for edge in missing_edges:
-    k_number = edge[0]
+    k_number = edge.strip()
     if k_number in seen_files:
         continue
 
-    print(f"{len(seen_files)} / {len(missing_edges)} ({round(len(seen_files) / len(missing_edges) * 100, 5)}%) Completed")
+    print(
+        f"{len(seen_files)} / {len(missing_edges)} ({round(len(seen_files) / len(missing_edges) * 100, 5)}%) Completed"
+    )
     nearest_percentage = math.ceil(len(seen_files) / len(missing_edges) * 100)
-    to_nearest_percentage = len(missing_edges) * math.ceil(len(seen_files) / len(missing_edges) * 100) // 100  - len(seen_files)
+    to_nearest_percentage = len(missing_edges) * math.ceil(
+        len(seen_files) / len(missing_edges) * 100
+    ) // 100 - len(seen_files)
     print(f"{to_nearest_percentage} more files to {nearest_percentage}% 🎉")
     if to_nearest_percentage == 0:
         print("🎉🎉🎉🎉🎉🎉🎉🎉🎉")
+
     process_pdf(k_number)
     seen_files.add(k_number)
